+++ src/DateTimeUtils.cpp
@@ -2,9 +2,9 @@
 
 #define LEAP_YEAR(Y) (((1970 + Y) > 0) && !((1970 + Y) % 4) && (((1970 + Y) % 100) || !((1970 + Y) % 400)))
 
-unsigned int daysInMonth(unsigned int year, unsigned int month)
+byte daysInMonth(unsigned int year, byte month)
 {
-    unsigned int nbDaysInMonth = 0;
+    byte nbDaysInMonth = 0;
     if (month == 2)
     {
         // compute number of days in the current month, according to the current year (leap or not)
